fix alerts chart empty state
